+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 * Allow configuring the `spimdisasm` section on extension segments (that inherit from `asm`, `data`, `rodata` or `bss` segments) before running the analisys on it.
   * This is done by overriding the `configure_disassembler_section` method.
 * New ps2-specific segments:
-  * `lit4` and `lit8`: Literal sections that only contain `float`s and `double`s respectively.
+  * `lit4` and `lit8`: "Literal" sections that only contain `float`s and `double`s respectively.
+  * `ctor`: Data pointing to C++ global data initialization functions.
 * `spimdisasm` 1.23.0 or above is now required.
 
 ### 0.22.3